Code review updates
Move Stop to src/main/scala/
