+++ core/src/main/scala-2/chisel3/VerificationStatement.scala
@@ -3,9 +3,6 @@
 package chisel3
 
 import chisel3.experimental.{BaseModule, SourceInfo}
-import chisel3.internal._
-import chisel3.internal.Builder.pushCommand
-import chisel3.layer.block
 
 import scala.language.experimental.macros
 
@@ -152,38 +149,4 @@ object cover extends VerifPrintMacrosDoc {
     macro VerifStmtMacrosCompat.cover._applyMacroWithMessage
   def apply(cond: Bool)(implicit sourceInfo: SourceInfo): Cover =
     macro VerifStmtMacrosCompat.cover._applyMacroWithNoMessage
-}
-
-object stop {
-
-  /** Named class for [[stop]]s. */
-  final class Stop private[chisel3] () extends VerificationStatement
-
-  /** Terminate execution, indicating success and printing a message.
-    *
-    * @param message a message describing why simulation was stopped
-    */
-  def apply(message: String)(implicit sourceInfo: SourceInfo): Stop = buildStopCommand(Some(PString(message)))
-
-  /** Terminate execution, indicating success and printing a message.
-    *
-    * @param message a printable describing why simulation was stopped
-    */
-  def apply(message: Printable)(implicit sourceInfo: SourceInfo): Stop = buildStopCommand(Some(message))
-
-  /** Terminate execution, indicating success.
-    */
-  def apply()(implicit sourceInfo: SourceInfo): Stop = buildStopCommand(None)
-
-  private def buildStopCommand(message: Option[Printable])(implicit sourceInfo: SourceInfo): Stop = {
-    val stopId = new Stop()
-    block(layers.Verification, skipIfAlreadyInBlock = true, skipIfLayersEnabled = true) {
-      message.foreach(Printable.checkScope(_))
-      when(!Module.reset.asBool) {
-        message.foreach(PrintfMacrosCompat.printfWithoutReset(_))
-        pushCommand(chisel3.internal.firrtl.ir.Stop(stopId, sourceInfo, Builder.forcedClock.ref, 0))
-      }
-    }
-    stopId
-  }
 }
